Changed MS classes for rownr_t

+++ ms/MSOper/MSSummary.h
@@ -110,7 +110,7 @@ public:
   ~MSSummary();
 
 // Retrieve number of rows
-   Int nrow() const;
+   Int64 nrow() const;
  
 // Retrieve image name
    String name() const;
